+++ git-remote-updater.php
@@ -14,7 +14,7 @@
  * Description:       Allows you to easily update Git Updater repositories in bulk via REST API endpoint updating. Requires Git Updater PRO.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           2.4.1
+ * Version:           2.4.2
  * License:           MIT
  * Network:           true
  * Domain Path:       /languages
@@ -38,19 +38,13 @@ if ( ! defined( 'WPINC' ) ) {
 // Load Autoloader.
 require_once __DIR__ . '/vendor/autoload.php';
 
-// Make sure `is_plugin_active()` is available.
-if ( ! function_exists( 'is_plugin_active' ) ) {
-	require_once ABSPATH . '/wp-admin/includes/plugin.php';
-}
-
-// Don't load if Git Updater not running.
-if ( ! is_plugin_active( 'git-updater/git-updater.php' ) ) {
-	return;
-}
-
 add_action(
 	'plugins_loaded',
 	function() {
+		// Don't load if Git Updater not running.
+		if ( ! is_plugin_active( 'git-updater/git-updater.php' ) ) {
+			return;
+		}
 		( new GRU_Freemius() )->init();
 		( new Bootstrap() )->run();
 	}