+++ src/Git_Remote_Updater/Webhooks.php
@@ -32,7 +32,7 @@ trait Webhooks {
 	/**
 	 * Holds data for repositories.
 	 *
-	 * @var \stdClass
+	 * @var array
 	 */
 	public $repos;